refact: vote count refact
